支持指定workerIdBits和sequenceBits

+++ src/main/java/butterfly/assigner/WorkerIdAssigner.java
@@ -7,5 +7,5 @@ package butterfly.assigner;
  */
 public interface WorkerIdAssigner {
 
-    long getWorkId();
+    int getWorkId();
 }
